Fix relocation unit test to reproduce Marco/Nadia bug
git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9569 0bcb941b-74d1-0310-bb26-e6636707f451

+++ toolfun/movieManagement/test/TestMovieData.m
@@ -25,7 +25,9 @@ classdef TestMovieData < TestCase
         
         function tearDown(self)
             delete(self.movie);
-            rmdir(self.moviePath,'s');
+            if isdir(self.moviePath)
+                rmdir(self.moviePath,'s');
+            end
         end
         
         %% Tests
